akra changes

+++ database/migrations/2017_03_25_060014_create_athlete_bioinformations_table.php
@@ -27,7 +27,7 @@ class CreateAthleteBioinformationsTable extends Migration
             $table->decimal('athlete_weight',10,3);
             $table->string('athlete_fathername',400);
             $table->string('athlete_passportNo',300);
-            $table->integer('athlete_cid');
+            $table->bigInteger('athlete_cid');
             $table->string('athlete_associatedSport',500);
             $table->string('athlete_photo',250);
             $table->integer('athlete_function')->default(2);
